Falcon-40B demo: load reference weights via native loader (fix transformers 5.x e2e hang, #47924)
The demo loaded its reference weights through the vendored
FalconForCausalLM.from_pretrained, which silently fails to populate the model
under transformers 5.x (from_pretrained reports missing_keys=0 but copies
nothing). The TT model, built from that empty state_dict, then hangs on device
during the first prefill — test_demo.py has timed out on the scheduled e2e
pipeline since the 5.10.2 bump (#47218). Bisected on T3000 e2e and confirmed.

Move the native-checkpoint loader from tests/test_utils.py into a shared
reference/load_falcon_weights.py (so non-test code can use it without importing
from tests/), have test_utils re-export it, and load the demo's reference model
through it. Same fix family as the unit-test fix in #47929.

Validated: Falcon-40B e2e demo PASSES with the native loader (was a 10-min hang).
